refactor: split snapshot generation into standalone script
Extracts generateSnapshots() into scripts/generate-snapshots.ts so it
can be run independently via npm run db:generate-snapshots without
re-importing CSV data. seed-e2e-data.ts imports and calls it.
